move log related decorators to decorators module
* usesFlushLoggedErrors
* usesFlushWarnings
