+++ master/buildbot/test/unit/test_config.py
@@ -31,9 +31,10 @@ from buildbot.process import factory
 from buildbot.process import properties
 from buildbot.schedulers import base as schedulers_base
 from buildbot.status import base as status_base
-from buildbot.test.util import compat
 from buildbot.test.util import dirs
 from buildbot.test.util.config import ConfigErrorsMixin
+from buildbot.test.util.decorators import usesFlushLoggedErrors
+from buildbot.test.util.decorators import usesFlushWarnings
 from buildbot.util import service
 from twisted.internet import defer
 from twisted.trial import unittest
@@ -236,7 +237,7 @@ class MasterConfig(ConfigErrorsMixin, dirs.DirsMixin, unittest.TestCase):
             lambda: config.MasterConfig.loadConfig(
                 self.basedir, self.filename))
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def test_loadConfig_parse_error(self):
         self.install_config_file('def x:\nbar')
         self.assertRaisesConfigError(
@@ -734,7 +735,7 @@ class MasterConfig_loaders(ConfigErrorsMixin, unittest.TestCase):
         self.assertIsInstance(self.cfg.builders[0], config.BuilderConfig)
         self.assertEqual(self.cfg.builders[0].name, 'x')
 
-    @compat.usesFlushWarnings
+    @usesFlushWarnings
     def test_load_builders_abs_builddir(self):
         bldr = dict(name='x', factory=factory.BuildFactory(), slavename='x',
                     builddir=os.path.abspath('.'))
@@ -1385,7 +1386,7 @@ class ReconfigurableServiceMixin(unittest.TestCase):
             self.assertEqual(prio_order, called_order)
         return d
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     @defer.inlineCallbacks
     def test_multiservice_nested_failure(self):
         svc = FakeMultiService()
