+++ master/buildbot/test/unit/test_mq_base.py
@@ -16,7 +16,7 @@
 import mock
 
 from buildbot.mq import base
-from buildbot.test.util import compat
+from buildbot.test.util.decorators import usesFlushLoggedErrors
 from twisted.internet import defer
 from twisted.python import failure
 from twisted.trial import unittest
@@ -41,7 +41,7 @@ class QueueRef(unittest.TestCase):
 
         cb.assert_called_with('rk', 'd')
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def test_exception(self):
         cb = mock.Mock(name='cb')
         cb.side_effect = RuntimeError('oh noes!')
@@ -52,7 +52,7 @@ class QueueRef(unittest.TestCase):
         cb.assert_called_with('rk', 'd')
         self.assertEqual(len(self.flushLoggedErrors(RuntimeError)), 1)
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def test_failure(self):
         cb = mock.Mock(name='cb')
         cb.return_value = defer.fail(failure.Failure(RuntimeError('oh noes!')))
