+++ master/buildbot/test/unit/test_status_web_hooks_gitorious.py
@@ -18,7 +18,7 @@ import calendar
 import buildbot.status.web.change_hook as change_hook
 
 from buildbot.test.fake.web import FakeRequest
-from buildbot.test.util import compat
+from buildbot.test.util.decorators import usesFlushLoggedErrors
 
 from twisted.trial import unittest
 
@@ -102,7 +102,7 @@ class TestChangeHookConfiguredWithGitChange(unittest.TestCase):
         d.addCallback(check_changes)
         return d
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def testGitWithNoJson(self):
         self.request = FakeRequest()
         self.request.uri = "/change_hook/gitorious"
