+++ master/buildbot/test/unit/test_steps_shell.py
@@ -28,9 +28,9 @@ from buildbot.steps import shell
 from buildbot.test.fake.remotecommand import Expect
 from buildbot.test.fake.remotecommand import ExpectRemoteRef
 from buildbot.test.fake.remotecommand import ExpectShell
-from buildbot.test.util import compat
 from buildbot.test.util import config as configmixin
 from buildbot.test.util import steps
+from buildbot.test.util.decorators import usesFlushLoggedErrors
 from twisted.trial import unittest
 
 
@@ -482,7 +482,7 @@ class SetPropertyFromCommand(steps.BuildStepMixin, unittest.TestCase):
                            state_string="'cmd' (failure)")
         return self.runStep()
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def test_run_extract_fn_exception(self):
         def extract_fn(rc, stdout, stderr):
             raise RuntimeError("oh noes")
@@ -853,7 +853,7 @@ class WarningCountingShellCommand(steps.BuildStepMixin, unittest.TestCase):
         return self.do_test_suppressions(step, supps_file, stdout, 2,
                                          exp_warning_log)
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def test_suppressions_warningExtractor_exc(self):
         def warningExtractor(step, line, match):
             raise RuntimeError("oh noes")
