+++ master/buildbot/test/unit/test_steps_slave.py
@@ -23,8 +23,8 @@ from buildbot.status.results import FAILURE
 from buildbot.status.results import SUCCESS
 from buildbot.steps import slave
 from buildbot.test.fake.remotecommand import Expect
-from buildbot.test.util import compat
 from buildbot.test.util import steps
+from buildbot.test.util.decorators import usesFlushLoggedErrors
 from twisted.internet import defer
 from twisted.trial import unittest
 
@@ -121,7 +121,7 @@ class TestFileExists(steps.BuildStepMixin, unittest.TestCase):
                            state_string="File not found. (failure)")
         return self.runStep()
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     @defer.inlineCallbacks
     def test_old_version(self):
         self.setupStep(slave.FileExists(file="x"),
