+++ master/buildbot/test/unit/test_steps_transfer.py
@@ -35,8 +35,8 @@ from buildbot.status.results import SUCCESS
 from buildbot.steps import transfer
 from buildbot.test.fake.remotecommand import Expect
 from buildbot.test.fake.remotecommand import ExpectRemoteRef
-from buildbot.test.util import compat
 from buildbot.test.util import steps
+from buildbot.test.util.decorators import usesFlushLoggedErrors
 from buildbot.util import json
 
 from cStringIO import StringIO
@@ -187,7 +187,7 @@ class TestFileUpload(steps.BuildStepMixin, unittest.TestCase):
         d = self.runStep()
         return d
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def testException(self):
         self.setupStep(
             transfer.FileUpload(slavesrc='srcfile', masterdest=self.destfile))
@@ -262,7 +262,7 @@ class TestDirectoryUpload(steps.BuildStepMixin, unittest.TestCase):
         d = self.runStep()
         return d
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def testException(self):
         self.setupStep(
             transfer.DirectoryUpload(slavesrc='srcdir', masterdest=self.destdir))
@@ -406,7 +406,7 @@ class TestMultipleFileUpload(steps.BuildStepMixin, unittest.TestCase):
         d = self.runStep()
         return d
 
-    @compat.usesFlushLoggedErrors
+    @usesFlushLoggedErrors
     def testException(self):
         self.setupStep(
             transfer.MultipleFileUpload(slavesrcs=["srcfile", "srcdir"], masterdest=self.destdir))
